hwbench: move to ruff format instead of black
ruff format has been out for while now, and is mostly compatible with
black. Only difference is the way it formats type hint comments (see
diff).

Be careful before merging this: it implies changing your editor config
(again, I know) to format things with ruff instead of black.

On the plus side, it does remove one (dev) dependency, and should be
faster (although this project is still a bit small to see the
improvements).

Don't forget, while the pre-commit in CI checks for formatting (and
fails if badly formatted), you can still use "make format" to do the
auto-formatting locally.