pex generation: remove use of lock file and wheels
The way we generate pex files broke when we added more dependencies (we
use numpy for graph generation). The choice was to build from source,
without wheels; but for that we need to be able to re-build numpy, and
this can be time-consuming.

If we remove the --no-wheel option, we go back to the exact versions
pinned in the generated requirements/base.txt; this follows the initial
intent of having reproducible builds for the pex. But this does not work
because the current pex implementation does not support locking versions
like this, while also having a local package (here hwbench itself) in a
directory (and not a wheel which has a hash). We get this error:

 ERROR: Can't verify hashes for these file:// requirements because they point to directories:

See this pex issues and the other linked ones:
https://github.com/pex-tool/pex/issues/429 ; currently the pex project
is building its own lock file format to better fit its needs, but that
would not help us here.

So this change relaxes the constraints a bit: no more lock file. The pex
builds won't be reproducible, and hopefully it won't be an issue since
the artifacts will be archived. And no more wheels: this should make the
pexes generation significantly faster.

Also upgrade the dependency resolver to be the same one we use with
pip-compile (backtracking), so that we'd get a similar result if pex and
requirements/base.txt are generated at the same time.

Change-Id: If6d677838f86880776a125c1f5676cbf1b08ddbb