hwbench: Renaming config file option
The -c/config option was used to point the file containing all jobs
while the BMC credentials were expected to be in an implicit file named
"config.cfg".

This generates unecessary confusion for users.
This commit is part of a series to clarify that.

This commit rename "-c/-config" to "-j/--jobs-config".
The pointed file must be a file containing the jobs to run.

This commit also add an "Examples" section in the README to help new
comers having a set of basic syntax to use for a first run.

Signed-off-by: Erwan Velu <e.velu@criteo.com>