hwbench: Add monitoring config file option
The existing code was using a hardcoded 'config.cfg' file to read the
BMC credentials in case of monitoring.

The hardcoded filename was not a great a idea, having a option to select
it would be clearly better.

The existing code was also referencing the credential file as 'config'
which could be misleading, let's rename it to be more explicit.

This commit:
- add -m/--monitoring-config option to select a file with credentials
- Add a sample "monitoring.cfg" credential file
- Add the -m usage in the Readme
- Add checks in connect_redfish() to ensure a file is selected and exist.
- Move self.bmc.connect_redfish() from Vendor.preapre() to Benchmarks.__schedule_benchmark()
  The existing code was calling connect_redfish() once the Vendor object
  is created. This was an issue since the BMC connection is always established even if
  the jobs file does not require any  monitoring.
  By moving this call, the connect_redfish() is called at the first time
  a job with monitoring is found. This ensure that if no monitoring is
  required, hwbench will not try to use the redfish credentials to login
  the BMC.

Signed-off-by: Erwan Velu <e.velu@criteo.com>