test: ensure every commit in a PR works
If a pull request contains multiple commits, it is advisable to ensure
that every single commit has been tested.

This modifies the python-check workflow to fetch all the commits in the
pull request and iteratively run the 'make check' command on each of
them. While we are at it, this slighly beautifies the output of the
github workflow.

Signed-off-by: Vincent Minet <v.minet@criteo.com>