hwbench/monitoring: Simplify addition of a new metric
The current code was a great set of code that got copy/paste over
various functions and vendors.

This generated lots of hard to read code which is error-prone.

This commit is merging all the metric addition logic in BMC.add_monitoring_value()
so it can be used across all vendors including the mocked one.

Signed-off-by: Erwan Velu <e.velu@criteo.com>