hwbench: Add PDU support
This commit add PDU as a type of monitoring device.

A monitoring.md documentation is added to detail how the monitoring
works and the way to use it.

PDU monitoring is optional and needs to be defined inside the monitoring
config file. It is possible to setup multiple PDUs configuration in the
same file.

This version implements a first driver for Raritan PDUs with Redfish
API.

At statup time the output looks like the following:

	Monitoring/PDU: initialize myPDUgroup with Raritan driver @ https://1url.pl/ Model: 'PX3-5722V-V2' FW: '4.2.0.5-50274' Serial: 'XXXXXXXXX'
	Monitoring/PDU: initialize myPDUdirect with Raritan driver @ https://url.pdu/ Model: 'PX3-5722V-V2' FW: '4.2.0.5-50274' Serial: 'XXXXXXXXX'
	Monitoring/BMC: Thermal metrics:1xCPU, 1xIntake
	Monitoring/BMC: Fans metrics:10xFan
	Monitoring/BMC: PowerConsumption metrics:65xCPU, 4xBMC
	Monitoring/BMC: PowerSupplies metrics:2xBMC
	Monitoring/PDU: PowerConsumption metrics:65xCPU, 4xBMC, 2xPDU

On the implementation side, the vendor drivers, like Raritan here, are
inheriting the PDU class and only override methods if necessary.

Signed-off-by: Erwan Velu <e.velu@criteo.com>