monitoring: Do not log error for all failed get_redfish_url()
The current implementation of get_redfish_url() logs an error every time
the get fails.

This was added to prevent silent failures.
Some endpoints must be present to get a proper monitoring.

That said, some calls may depends on the hardware specification meaning
that several calls can be needed to get the proper metrics.
In that case, the first call will complains while the second will work.
This generates very noisy output that makes users feeling something is
wrong while everything works as expected.

This commit is adding a log_failure paramater to get_redfish_url() to
indicate if the call should log a failure or not. This is up to the
caller to determine if this flag is necessary.

The current implementation keeps the same behavior aka logging every error.

This commit silence two calls
- one from HPE to detect if we run into a multinode chassis
- one from Dell where the OEM metrics can be located in two different paths.

Signed-off-by: Erwan Velu <e.velu@criteo.com>