dell: Do not discover OEM endpoint at every run
get_oem_system() is trying to find the right endpoint as depending on
systems they might vary.

During the monitoring loop, in your are in the case where we need to use
the /Attributes, the function generates two redfish calls.

Once the right endpoint is found, let's reuse it directly at each call
to save time and reduce the number of redfish call we perform.

Signed-off-by: Erwan Velu <e.velu@criteo.com>