hwgraph: Adding metric get_full_name()
When loading metrics, it appear that some can have the same name making
invalid data, data name and misleading graphs.

This was seen in issue #36.

The BMC reported the two following structures :

          "CPU": {
            "02-CPU 1 PkgTmp": {
              "name": "CPU1",
              "unit": "Celsius",
              "mean": [
	  [...]
            "12-CPU 1": {
              "name": "CPU1",
              "unit": "Celsius",
              "mean": [

As we use the name member, we got the two different metrics being merged
as a single one.

This patch is adding a workaround by consider the parent's name as the
"full_name" of the metric.

So this patch is adding this value, like "12-CPU 1", as the full name
and reachable via get_full_name().

This commit ensures the env graphs are using this full name to avoid
smashing metrics.

Signed-off-by: Erwan Velu <e.velu@criteo.com>