hwbench: Using absolute paths for out & tuning directories
As per issue #38, the dmidecode call to save the SMBIOS table in a
binary file for later reading is broken as the file is not present into
the hwbench-out directory.

It appear that when dmidecode --dump-file is called, that is done with a
relative path.

The stderr log show the following error :
	hwbench-out-20241025123459/dmidecode.bin: open: No such file or directory

It appear the relative path passed to dmidecode is not existing because
dmidecode is not started in the same home dir as we are currently
running the tool.

This commit is about using an absolute path instead of a relative one to
ensure that every command/tool can dump content in the proper directory.

With this patch, the dmidecode-bin file is now present in the
hwbench-out archive.

Signed-off-by: Erwan Velu <e.velu@criteo.com>