hwgraph: Add option to manage missing data points.
As reported in issue #40, when data points are missing,
hwgraph was crashing because graphs must have the same
number of data points.

In such case, let's default to a fatal message to the user instead of
crahsing.

This commit also add a new hwgraph option, --ignore-missing-datapoint
that can be used to define a behavior to ignore this case and graph.

  --ignore-missing-datapoint=zero will replace the missing datapoint by a 0 value.
  --ignore-missing-datapoint=last will replace the missing datapoint by the last known value of the serie.

It's up to the user to decide which one is the best but also understand
that graphs will be partially wrong.

By default, hwgraph generates a fatal error.

Signed-off-by: Erwan Velu <e.velu@criteo.com>