graph: Adding CPU clock scaling graph
As per issue #39, it could be super useful to have a new type of graph
to represent how the CPU clock performed during a benchmark.

This new graph gives a quick overview on how the CPU frequency behave
during a scaling benchmark.

The rendering is not as precise as environment graphs, but it gives a
brief overview with the following trade-offs for the y-err bars:
- min of the yerr-bar, is the min of min values
- mean of the yerr-bar, is the mean of mean values
- max of the yerr-bar, is the max of the max values

Signed-off-by: Erwan Velu <e.velu@criteo.com>