bmc: add default implementation for thermal, fans and power
This fetches all available chassis in the redfish tree by default, and
iterates over sub-elements.

The list of chassis, and thermal or power URIs are cached globally, so
they are only fetched once per run.

The current implementation will only work for fans and power if there is
only one chassis element with "Thermal" or "Power" information.

This is because we keep the default implementation of get_power() and
get_thermal() in order not to break the API for other vendors for fans
and power consumption.