hwbench: generic vendor: make monitoring file optional
Since we moved to the generic vendor, we no longer use the mockedvendor
by default (it's now only for tests). So when running hwbench on a
desktop/laptop for local testing, it requires setting the monitoring
file, which wasn't the case before.

This change makes the monitoring file optional in this case, with a
fallback when it is not provided to the MockedBMC.