stressng: fix results parsing on big servers
In some cases, one might have big numbers in the stress-ng output, which
will reduce the number of spaces, leading to having only one space ' '
between different output elements. But the regex is designed to have at
least two spaces between elements, so fix it to bring it to a minimum of
1.

This adds a test that shows the failing stress-ng parsing, and adds the
fix at the same time.