hwbench: Dump /sys/block at startup
When hwbench starts, it's interesting to collect the block device
status.

Having the list of devices that were detected with their associate
state is interesting to have when considering the behavior of a
particular host : i.e having idle rotational drives influence the
overall power consumption.

This patch lists all block devices from /sys/block and make a specific
tarball with it.

A typical output with a machine having 6 nvme and 24 rotational drives :

	[root@88-e9-a4-d7-a6-a4 hwbench]# ls hwbench-out-20250502094534/ | grep block
	sys-block-nvme0n1.tar
	sys-block-nvme1n1.tar
	sys-block-nvme2n1.tar
	sys-block-nvme3n1.tar
	sys-block-nvme4n1.tar
	sys-block-nvme5n1.tar
	sys-block-sda.tar
	sys-block-sdb.tar
	sys-block-sdc.tar
	sys-block-sdd.tar
	sys-block-sde.tar
	sys-block-sdf.tar
	sys-block-sdg.tar
	sys-block-sdh.tar
	sys-block-sdi.tar
	sys-block-sdj.tar
	sys-block-sdk.tar
	sys-block-sdl.tar
	sys-block-sdm.tar
	sys-block-sdn.tar
	sys-block-sdo.tar
	sys-block-sdp.tar
	sys-block-sdq.tar
	sys-block-sdr.tar
	sys-block-sds.tar
	sys-block-sdt.tar
	sys-block-sdu.tar
	sys-block-sdv.tar
	sys-block-sdw.tar
	sys-block-sdx.tar
	[root@88-e9-a4-d7-a6-a4 hwbench]#

Signed-off-by: Erwan Velu <e.velu@criteo.com>