hwgraph/spike: Print cycles when using manual spiking
When a manual spike configuration is used, like 10s high, 5s low,
the current code was not displaying any message during the run.

This was a bit confusing at it could look like hwbench has an issue.

This commit add a message for each iteration to show steps and
their duration.

Signed-off-by: Erwan Velu <e.velu@criteo.com>