cpu_cores: make sure to properly select hyperthread cores
With the new sorting from the previous commit, it should support more situations, such as the case of VMs with Hyper-V.
More technically speaking, it is enumerating cores sequentially:
```
lscpu --all --parse=SOCKET,CORE,CPU
0,0,0
0,0,1
0,1,2
0,1,3
0,2,4
0,2,5
0,3,6
0,3,7
0,4,8
0,4,9
0,5,10
0,5,11
```

Which was breaking the previous code expecting something else. Now we don't expect anything, we just sort and exclude the first CPU on the core, expecting all the others to be hyperthread cores (could be more than one).