ux: crash early if a binary is not available, and display all the missing ones
This should make the entry into the hwbench world easier. Now, instead of crashing once and for all if something is missing, it will collect everything missing, and then stop. This system can be easily expanded to add other kind of requirements.

For instance, it returns this when I'm missing those binaries:
```
CRITICAL:root:Requirements are not met: Binary smartctl is not available on the system
CRITICAL:root:Requirements are not met: Binary sdparm is not available on the system
CRITICAL:root:Requirements are not met: Binary stress-ng is not available on the system
CRITICAL:root:Requirements are not met: Binary fio is not available on the system
```