hwgraph: Reduce dot size on x2 axis
When plotting large graphs with crossing or nearly parallel lines, using
the "o" marker makes a too large marker on the line. As a consequence,
the graph is hard to read and some part of the lines are nearly
unreadable.

This commit is using the smallest marker which is large enough to be
seen, increasing th readbility of the graphs.

Signed-off-by: Erwan Velu <e.velu@criteo.com>