hwgraph: Add option to graph events
Some benchmarks are operated with external scenario like:
- shutdown a PSU during a period of time
- remove liquid cooling during a portion of a run
- remove a fan
....

While rendering the graphs, it was not yet possible to indicate specific
period of the benchmark where some external events occured.

This commit adds a --events option to hwgraph to add to the graph some
visual indications that events occured at some period of the benchmark.

Each event is declared like <event_name>:<start_time>:<duration>, a
example looks like the following :
 uv run hwgraph graph
   --traces mydir/results.json:my_server_name:BMC.Server
   --title 'Full CPU load 10 minutes - no coolant for 90 secs'
   --outdir mydir/graphs
   --events no-coolant:0:90

In this example, a background nearly transparent box will be added for
the first 90 seconds (start_time=0s, duration=90s) of the benchmark.
A new entry will be added to the legend : "Event no-coolant"

As per the --traces option, multiple events can be defined, each new
event will be colored differently to ensure an easy reading.

Signed-off-by: Erwan Velu <e.velu@criteo.com>