hwgraph: Adding --same-group option
When performing full rack benchmarking, it could be interesting to graph
some specific graphs for studying the following metrics during the full
run:
- the sum of the servers power consumption
- the sum of the cpus power consumption
- the sum of difference between server's power and cpu power
  consumptions

This patch also prints the percentage ratio of the cpu power and the
other components versus the server's power consumption.
It could be interesting to see how much of a server power is consumed by the
CPUs.

This is about the same idea as the --same-chassis but at the scale of a
group of servers.

This feature, will be useful to graph results of a given rack.

Signed-off-by: Erwan Velu <e.velu@criteo.com>