hwgraph: Disable "Max versus" mode if not enough traces
The max_versus mode is useful to compare the same run over several
machines or the same machine in various conditions.

If a single trace is provided, the produced graphs are not very
interesting to read.
Let's disable them and print a message to explain the reason of
this automatic feature disabling.

Signed-off-by: Erwan Velu <e.velu@criteo.com>