hwbench: pdus: multiple outlets per PDU
This allows splitting the outlet or outletgroup options with a comma,
and summing the measures to be able to create a measure.

A new PDU option 'separator' can be used to change the default from a
comma to an arbitrary string.

This is useful if you can't create outletgroups or don't want to modify
its config to create outletgroups, or want to aggregate multiple groups
of the same PDU.