hwgraph: Increase Percent's scale to 110
When rendering units in percent, using 100% is problematic since this
will be the maximum value of the graph.

If the measured value is 100%, the graph will mix data and the frame of
the graph making the data unreadable.

This commit increase the max value to 110, so a 100% will be graphed
below the max possible value, making it readable.

Signed-off-by: Erwan Velu <e.velu@criteo.com>