hwbench: fix monitoring loop when late
This should fix the monitoring loop when we are late and make the code a
bit easier to read.

Before this change, the code would loop indefinitely if an iteration of
the monitoring was late by more than 500ms.

Check that there is still time to run based on the current time, not a
calculation based on the current loop runtime. And return if there is no
time, ignoring the negative sleep time.