hwbench: turbostat: initialize and run turbostat in parallel
We noticed when running hwbench on a big CPU, Turbostat was really slow to initialize. Unfortunately, we reran it every iteration of the monitoring loop. The metrics were constantly late and could not be used.

Now, we run Turbostat at the beginning and synchronizing the monitoring loop. The system shall take the closest Turbostat output, to make sure it's overlapping the most in term of timestamps.
And with that, Turbostat can take the time it needs to initialize, we shall be good for metrics!