chore(monitoring): refactor monitoring types
Until today, the monitoring was kind of a mess of imbricated dicts. This was working fine, of course, and there was enums everywhere to make sure the keys are working okay.
But this was extremely hard to read and follow, and you never knew at which level you were before today.
Also, the type checkers were all losts, and "Unknown" was the norm.

Now the fields are all defined, the types are correct and detected. This was not as easy as expected, and some methods still returns "Any" on purpose (notably "getting" a metric) since Python do not support a "keyof" kind of operator.
There was some regressions, such as in the graphing part where some keys are now hardcoded. FAR, we should implement the proper MonitoringData struct and serialize (that should not be too hard) to make sure that everything is ISO. This may make backward compatibility harder, though.

That said, today, there should NOT be any incompatibilities with older traces (at least, as of before this commit) even if the tree slightly changed.