fix(turbostat): get metrics when asked
We noticed that Turbostat have a weird behavior when waiting for metrics and our implementation had some difficulties with recent versions of Python.
We also found out that Turbostat flushes metrics after sending an EOL so this commit removes most of the time sensitive code and simply flushes in each iteration asynchronously.

This fixes all the sync issues as well as simplify the code quite a lot. All of this while having an even better synchronization than before.
Aside, it does remove the need for the debug flag.